requirements.txt
Django==4.2.11
python-dotenv==1.2.1
Pillow==10.2.0